refactor: Typography 폴더로 Text 컴포넌트 이동
