+++ src/components/Common/Typography/Text/Text.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
 import Text from './Text';
-import { sizes } from './text.types';
+import { sizes } from '../typography.types';
 
 const meta: Meta<typeof Text> = {
   title: 'common/Text',
